move #nav outside of main content
This has the effect of excluding it from the TOC, which will save space.
This is only for pages that use template.html, so index.html is not affected.
Still TODO: implement a separate nav() command to show these links.

tui: also set display: none on #nav during boot() and display: block
during shutdown() so it doesn't show up at the top of the page