tui: fix a bug related to uppercase node names
I noticed this problem when I added a node name ("Makefile")
containing an uppercase letter and tried to display it using the
naming shortcut.

The problem was in sh():  it was using the lowercased version of its
first argument ("cmd") as the node ID instead of the string as typed
by the user. The fix is to use the string directly to search for the node.