tui: save shutdown state in localStorage for persistence across pages
This required a little shifting around of the code and careful
management of global state:

Since boot() is responsible both for initializing the global variables
and displaying the terminal interface, make sure everything gets
initialized first, then check the shutdown state, and only hide the
static data at the last second.

Add a reboot() click handler to clear the shutdown state in
localStorage before calling boot() again.

Since shutdown() is both responsible for undoing the changes by boot()
and displaying the message to reboot, it's necessary to switch on
uiRoot.display (which is set to none in the CSS) explicitly, for the
case where shutdown() is called by boot() due to a shutdown on a
previous page.  Otherwise the reboot message won't appear for the user.